move act -> page
